Refactor to support MM-evals. (#3)
This PR refactors the code originally only created for MM-MT-Bench and instead supports a generic set of evals. This is meant to be lightweight and extensible.

We release prompt + metrics + eval runner for the minimal set of evals reported in the Pixtral-12B paper, and welcome contributions for more evals in a standardized format.

    DocVQA
    MMMU
    MathVista
    ChartQA

MM-MT-Bench is also included from before.